+++ src/Networking/NexusMods.Networking.Downloaders/Tasks/NxmDownloadTask.cs
